# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spillway admin API
  version: 0.4.0
  description: |
    The API behind the dashboard. Its only caller is the dashboard's server-side layer, which signs users in
    with POST /admin/login and forwards the returned token as `Authorization: Bearer <token>`. The browser
    never sees the token. TypeScript types for web/ are generated from this file, and a Go test checks that
    the real responses match it.

    Each operation says which roles may call it with `x-access`: `public`, `viewer` (any signed-in user)
    or `admin`. The Go router registers the same access level, and a test fails if they disagree.
paths:
...
  /admin/status:
    get:
      operationId: status
      summary: Whether Spillway's dependencies are reachable
      x-access: viewer
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Dependency status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "401":
          description: No token, a bad token, or an expired one
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /admin/runs/summary:
    get:
      operationId: runsSummary
      summary: Counts by status, and the runs waiting for a person
      description: Running, Sleeping and Needs you are the runs in those states now; Succeeded, Failed and Cancelled are runs that ended inside the range.
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: hours, in: query, required: false, schema: { type: integer, enum: [1, 24, 168], default: 24 } }
      responses:
        "200":
          description: Counts and the waiting list
          content:
            application/json:
              schema:
                type: object
                required: [hours, counts, waiting]
                properties:
                  hours: { type: integer }
                  counts: { $ref: "#/components/schemas/RunCounts" }
                  waiting:
                    type: array
                    items: { $ref: "#/components/schemas/WaitingRun" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /admin/runs/activity:
    get:
      operationId: runsActivity
      summary: Runs started per slice of time, by where they stand now
      description: One slice per 5 minutes for an hour, per hour for a day, per day for a week. Slices with no runs are included.
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: hours, in: query, required: false, schema: { type: integer, enum: [1, 24, 168], default: 24 } }
      responses:
        "200":
          description: Buckets, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [hours, bucket_seconds, buckets]
                properties:
                  hours: { type: integer }
                  bucket_seconds: { type: integer }
                  buckets:
                    type: array
                    items: { $ref: "#/components/schemas/ActivityBucket" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /admin/runs:
    get:
      operationId: runsList
      summary: Runs, newest first, each with a strip of its latest steps
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: state, in: query, required: false, description: "active is every run not finished; finished is runs that ended inside `hours`", schema: { type: string, enum: [active, finished] } }
        - { name: status, in: query, required: false, schema: { type: string, enum: [queued, running, waiting_tool, waiting_human, sleeping, succeeded, failed, cancelled] } }
        - { name: hours, in: query, required: false, schema: { type: integer, enum: [1, 24, 168], default: 24 } }
        - { name: q, in: query, required: false, description: Only runs whose task contains this text (ignoring case), schema: { type: string, maxLength: 100 } }
        - { name: key_id, in: query, required: false, description: Only runs of this API key, schema: { type: string, format: uuid } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: cursor, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: One page
          content:
            application/json:
              schema:
                type: object
                required: [runs, next_cursor]
                properties:
                  runs:
                    type: array
                    items: { $ref: "#/components/schemas/RunItem" }
                  next_cursor: { type: string, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: runsCreate
      summary: Start a run from the dashboard
      description: The run is made under the built-in playground key and charged to it. Tools are not available from the dashboard yet.
      x-access: admin
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RunStart" }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/runs/{id}:
    get:
      operationId: runGet
      summary: One run, in the shape of a list row
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunItem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/runs/{id}/graph:
    get:
      operationId: runGraph
      summary: The run as a graph of attempts, with workers and recoveries
      description: One node per attempt at a step, so a step re-issued after a worker was lost appears once per attempt.
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The graph
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunGraph" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/runs/{id}/events:
    get:
      operationId: runEvents
      summary: The run's step rows as server-sent events
      description: Replays rows after Last-Event-ID (or ?after=), then streams live ones, and ends after the final run.status event.
      x-access: viewer
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: after, in: query, required: false, schema: { type: integer, format: int64, minimum: 0 } }
      responses:
        "200":
          description: An event stream
          content:
            text/event-stream:
              schema: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/runs/{id}/cancel:
    post:
      operationId: runCancel
      summary: Cancel a run
      description: A run nobody is working on is cancelled at once; a run a worker holds is cancelled within a heartbeat.
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "202":
          description: Cancelled, or asked to cancel
          content:
            application/json:
              schema:
                type: object
                required: [id, status, cancel_requested]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
                  cancel_requested: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The run already finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/runs/{id}/approve:
    post:
      operationId: runApprove
      summary: Approve a run that waits for a person
      description: Resolves the run's pending wait_human step. The run is claimable at once, by any worker, including one that starts after the worker that parked it has gone.
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Decision" }
      responses:
        "202": { $ref: "#/components/responses/Decided" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The run is not waiting for a decision, or already finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/runs/{id}/reject:
    post:
      operationId: runReject
      summary: Reject a run that waits for a person
      description: Ends the run as failed with reason "rejected". The note is kept on the step.
      x-access: admin
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Decision" }
      responses:
        "202": { $ref: "#/components/responses/Decided" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The run is not waiting for a decision, or already finished
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
...
  responses:
    Decided:
      description: The decision was recorded
      content:
        application/json:
          schema:
            type: object
            required: [id, status]
            properties:
              id: { type: string, format: uuid }
              status: { type: string }
    BadRequest:
      description: The request is not valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No token, a bad token, or an expired one
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The signed-in user is a viewer and this needs an admin
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
...
    Status:
      type: object
      required: [postgres, redis]
      properties:
        postgres: { type: string, enum: [up, down, unknown] }
        redis: { type: string, enum: [up, down, disabled], description: "disabled means REDIS_URL is not set" }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
    RunCounts:
      type: object
      required: [running, sleeping, needs_you, succeeded, failed, cancelled]
      properties:
        running: { type: integer }
        sleeping: { type: integer }
        needs_you: { type: integer }
        succeeded: { type: integer }
        failed: { type: integer }
        cancelled: { type: integer }
    WaitingRun:
      type: object
      required: [id, goal, tool, waiting_since, key]
      properties:
        id: { type: string, format: uuid }
        goal: { type: string }
        tool: { type: string }
        waiting_since: { type: string, format: date-time }
        key: { type: string, description: The name of the API key that started the run }
    ActivityBucket:
      type: object
      required: [start, succeeded, failed, cancelled, in_progress]
      properties:
        start: { type: string, format: date-time }
        succeeded: { type: integer }
        failed: { type: integer }
        cancelled: { type: integer }
        in_progress: { type: integer }
    StripNode:
      type: object
      required: [kind, state]
      properties:
        kind: { type: string, enum: [model, tool] }
        state: { type: string, enum: [done, current, failed, waiting] }
    Strip:
      type: object
      required: [steps, more]
      properties:
        steps: { type: array, items: { $ref: "#/components/schemas/StripNode" }, description: "At most 16: the latest ones" }
        more: { type: integer, description: How many earlier steps are not shown }
    RunItem:
      type: object
      required: [id, status, goal, key, model, step_count, cost_usd, created_at, finished_at, duration_ms, wake_at, failure_reason, strip]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, waiting_tool, waiting_human, sleeping, succeeded, failed, cancelled] }
        goal: { type: string }
        key: { type: string }
        model: { type: string }
        step_count: { type: integer }
        cost_usd: { type: string }
        created_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time, nullable: true }
        duration_ms: { type: integer, format: int64 }
        wake_at: { type: string, format: date-time, nullable: true }
        failure_reason: { type: string, nullable: true }
        strip: { $ref: "#/components/schemas/Strip" }
    RunStart:
      type: object
      required: [input]
      additionalProperties: false
      properties:
        input: { type: string }
        system: { type: string }
        model: { type: string, description: A policy name or model id; default is the default policy }
        limits:
          type: object
          additionalProperties: false
          properties:
            max_steps: { type: integer, minimum: 1 }
            max_cost_usd: { type: number }
            deadline_seconds: { type: integer, minimum: 1 }
        metadata: { type: object }
    GraphAttempt:
      type: object
      required: [provider, model, kind, latency_ms]
      properties:
        provider: { type: string }
        model: { type: string }
        kind: { type: string }
        latency_ms: { type: integer }
        status: { type: integer }
        error: { type: string }
        error_kind: { type: string }
        injected: { type: boolean }
        estimated: { type: boolean }
    Decision:
      type: object
      additionalProperties: false
      properties:
        note: { type: string, maxLength: 2000 }
...
    GraphNode:
      type: object
      required: [step_no, type, state, worker, epoch, reissued, started_at, duration_ms, cost_usd, attempts]
      properties:
        step_no: { type: integer }
        type: { type: string, enum: [model_call, tool_call, wait_human, sleep, compaction] }
        state: { type: string, enum: [finished, failed, running, stopped, waiting, sleeping] }
        worker: { type: string }
        epoch: { type: integer, format: int64 }
        reissued: { type: boolean }
        previous_worker: { type: string }
        previous_epoch: { type: integer, format: int64 }
        started_at: { type: string, format: date-time }
        duration_ms: { type: integer, format: int64, nullable: true, description: Null for an attempt that stopped }
        cost_usd: { type: string }
        tool: { type: string }
        model: { type: string }
        provider: { type: string }
        cache: { type: string }
        tokens:
          type: object
          required: [in, out]
          properties: { in: { type: integer }, out: { type: integer } }
        idempotency_key: { type: string }
        attempts: { type: array, items: { $ref: "#/components/schemas/GraphAttempt" } }
        arguments: { type: string }
        result: { type: string }
        message: { type: string }
        error: { type: string }
        reason: { type: string, description: "wait_human: why the run waits" }
        gate: { type: boolean, description: "wait_human: true when a tool is held for approval, false when the model asked" }
        decision: { type: string, enum: [approve, reject] }
        by: { type: string }
        note: { type: string }
        seconds: { type: integer, description: "sleep: how long" }
        wake_at: { type: string, format: date-time, description: "sleep: when it ends" }
    RunGraph:
      type: object
      required: [run, workers, nodes, recoveries, last_event_id]
      properties:
        run:
          type: object
          required: [id, status, goal, key, model, tools, step_count, cost_usd, max_steps, max_cost_usd, deadline_seconds, created_at, finished_at, deadline_at, failure_reason, cancel_requested, lease_owner, lease_epoch, lease_expires_at, wake_at, approval]
          properties:
            id: { type: string, format: uuid }
            status: { type: string, enum: [queued, running, waiting_tool, waiting_human, sleeping, succeeded, failed, cancelled] }
            goal: { type: string }
            key: { type: string }
            model: { type: string }
            tools: { type: array, items: { type: string } }
            step_count: { type: integer }
            cost_usd: { type: string }
            max_steps: { type: integer }
            max_cost_usd: { type: string }
            deadline_seconds: { type: integer }
            created_at: { type: string, format: date-time }
            finished_at: { type: string, format: date-time, nullable: true }
            deadline_at: { type: string, format: date-time }
            failure_reason: { type: string, nullable: true }
            cancel_requested: { type: boolean }
            lease_owner: { type: string, nullable: true }
            lease_epoch: { type: integer, format: int64 }
            lease_expires_at: { type: string, format: date-time, nullable: true }
            wake_at: { type: string, format: date-time, nullable: true, description: Set while the run sleeps }
            approval:
              nullable: true
              description: Set while the run waits for a person
              type: object
              required: [step_no, reason, gate, since]
              properties:
                step_no: { type: integer }
                reason: { type: string }
                tool: { type: string }
                arguments: { type: string }
                gate: { type: boolean }
                since: { type: string, format: date-time }
        workers:
          type: array
          items:
            type: object
            required: [id, epochs]
            properties:
              id: { type: string }
              epochs: { type: array, items: { type: integer, format: int64 } }
        nodes: { type: array, items: { $ref: "#/components/schemas/GraphNode" } }
        recoveries:
          type: array
          items:
            type: object
            required: [after_step, from_worker, to_worker, epoch, at]
            properties:
              after_step: { type: integer }
              from_worker: { type: string }
              to_worker: { type: string }
              epoch: { type: integer, format: int64 }
              at: { type: string, format: date-time }
        last_event_id: { type: integer, format: int64, description: The id of the newest step row; events after it are new }